Mancala

Documentation

data MancalaBoard

The MancalaBoard data type represents the Game Board. It is actually a list of BoardSides, that are representing the amount of stones for each player. CurrentPlayer represents the current player.

Constructors

MkMancalaBoard CurrentPlayer [BoardSide]

data Player

initMancalaBoard :: MancalaBoard

Overriding standard show to present the Mancala Board in some better way. |TODO: Zastanowic sie nad ladnym wyswietlaniem instance Show MancalaBoard where show (MkMancalaBoard stones) = show(stones)

Initial placing stones on Board. Na razie brak mozliwosci wyboru gracza. Na razie gre zaczyna zawsze gracz A. (TODO)

switchPlayer :: Player -> [Player] -> Player

Function that returns the next player from the queue.

initPlayerList :: [Player]

Function that initializes players list for mancala.

getBoardSideForPlayer :: MancalaBoard -> Player -> BoardSide

Function that returns BoardSide for a player

getPointsForPlayer :: MancalaBoard -> Player -> Int

Function that returns current score for the player. It is the amount of stones in the house, which belongs to that person, plus amount of stones in every pit on the side of their board.

checkIfMovePossible :: MancalaBoard -> Int -> Bool

Function that checks if wished move is possible.

makeMove :: MancalaBoard -> Int -> MancalaBoard

Function which takes a MancalaBoard and an integer representing the number of pit, from which stones are picked. If number of stones on the chosen pit is equal to (numberOfPits - pitNumber) + 14k for some integer k, than the move will end in the house of the current player.